Move Lightbox component to its own folder and fix height
